modify car model

+++ db/migrate/20230104160051_create_cars.rb
@@ -4,8 +4,9 @@ class CreateCars < ActiveRecord::Migration[7.0]
       t.string :name
       t.string :image
       t.string :model
-      t.float :daily_price
+      t.decimal :daily_price
       t.text :description 
+  
       t.timestamps
     end
   end
